resolved price expected type error, now checkout api data coming successfully

+++ app/src/main/java/com/softaai/footballmatchbooking/data/model/checkout/Checkout.kt
@@ -21,5 +21,5 @@ data class Checkout(
     @Json(name = "paymentMethodUnsupported")
     val paymentMethodUnsupported: String,
     @Json(name = "price")
-    val price: Int
+    val price: Float
 )
